- hopefully implemented opcode DXYN (drawing a sprite on screen) - added a method for reading multiple consecutive bytes from memory - decided to keep smaller data types for address registers and memory

+++ src/main/java/emulator/AddressRegister.java
@@ -8,13 +8,13 @@ package emulator;
 
 final class AddressRegister {
 
-    private int memory;
+    private short memory;
 
-    int read() {
+    short read() {
         return memory;
     }
 
-    void write(int value) {
+    void write(short value) {
         memory = value;
     }
